Time: 39 ms (63.76%), Space: 27.6 MB (69.64%) - LeetHub

+++ 704-binary-search/704-binary-search.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int search(vector<int>& nums, int target) {
-        long long n,s,e,mid,idx;
+        int n,s,e,mid,idx;
         n = nums.size();
         
         s = 0, e = (n-1), idx = 0;
